% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_version.R
\name{snapshot_versioned}
\alias{snapshot_versioned}
\title{Snapshot a versioned lockfile}
\usage{
snapshot_versioned(
  name,
  overwrite = FALSE,
  r_version = R.version,
  date = Sys.Date(),
  pkgs = pkg_list()
)
}
\description{
Snapshot to a specific lockfile, for later use as a share baseline.
}
\examples{
snapshot_version
}
